Add LCD screen (#4)
This removes the seven-segment display and adds a 16x2 character LCD screen. Messages are shown when the box initializes and receives updates, and always shows the current preset ID.

We don't currently save synth names on the box, so we can't display those until we start sending them over the wire and persisting them into EEPROM storage. In the meantime, the box just shows `1:[Preset Name]`.